requirements.txt
flwr==1.2.0
tensorflow==2.11.0
hydra-core==1.3.2
numpy==1.23.0
torch==1.13.1
torchvision==0.14.1
grpcio==1.51.1
matplotlib